luerl_old (luerl v1.5.1)

Legacy API for Luerl.

This module is provided for backwards compatibility with earlier versions of Luerl. New code should use the luerl module instead.

For Elixir users, the Elixir.Luerl.Old module provides an idiomatic interface with state as the first argument for better pipe operator usage.

Functions

call(C, As, St)

call_chunk(C, As, St0)

call_function1/3

call_function(Fp, As, St0)

call_method1(Fp, Las, St0)

call_method(Fp, As, St0)

clear_trace_func(St)

decode(LT, St)

decode_list(Lts, St)

do/2

dofile(Path, St0)

encode/2

encode_list(Ts, St)

eval(Chunk, St0)

evalfile(Path, St0)

externalize(S)

function_list(Ks, St)

gc(St)

get_stacktrace/1

get_table1/2

get_table/2

get_trace_data(St)

init()

internalize(S)

load(Bin, St)

load/3

load_module1(Lfp, Mod, St0)

load_module/3

loadfile(Name, St)

loadfile(Name, Opts, St0)

method_list/2

path_loadfile(Name, St)

path_loadfile(Dirs, Name, St)

path_loadfile/4

set_table1(Lfp, Lv, St)

set_table1(Tab, Key, Lv, St)

set_table/3

set_trace_data(Tdata, St)

set_trace_func(Tfunc, St)

stop(St)